requirements.txt
requests
Pillow
piexif
face_recognition
numpy
faiss-cpu
tqdm
python-dateutil
jupyter